refactor(client): flatten payment components directory
