+++ client/src/components/pages/Store/Wallet/Transactions/Payment/index.js
@@ -1,6 +1,6 @@
-export { PaymentTab } from "./components/PaymentTab";
-export { PaymentConfirmModal } from "./components/PaymentConfirmModal";
-export { PaymentSuccessContent } from "./components/PaymentSuccessContent";
+export { PaymentTab } from "./PaymentTab";
+export { PaymentConfirmModal } from "./PaymentConfirmModal";
+export { PaymentSuccessContent } from "./PaymentSuccessContent";
 export { getPaymentErrorDescription, PAYMENT_ERROR_TRANSLATIONS } from "./utils/paymentErrors";
 export { usePaymentAmountInput } from "./hooks/usePaymentAmountInput";
 export { useSatsToFiatEstimate } from "./hooks/useSatsToFiatEstimate";
